requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-openai
langchain-community

tortoise-orm
aiomysql
aerich
pypika-tortoise==0.2.2

chromadb
tiktoken

passlib
python-jose
dashscope
duckduckgo-search
bcrypt==5.0.0
